style: rename search related variables + move procedure
