+++ apps/web/src/types/entities.ts
@@ -59,6 +59,6 @@ export type DailyStatName = keyof Omit<
 
 export type OverallStats = AppRouterOutputs["stats"]["getOverallStats"][number];
 
-export type SearchOutput = NonNullable<AppRouterOutputs["search"]["byTerm"]>;
+export type SearchOutput = NonNullable<AppRouterOutputs["search"]>;
 
 export type SearchCategory = keyof SearchOutput;
