--- packages/api/src/routers/search/index.ts
@@ -1,6 +1,0 @@
-import { t } from "../../trpc-client";
-import { byTerm } from "./byTerm";
-
-export const searchRouter = t.router({
-  byTerm,
-});
