+++ packages/api/test/search.test.ts
@@ -1,11 +1,11 @@
 import type { inferProcedureInput } from "@trpc/server";
 import { beforeAll, describe, expect, it } from "vitest";
 
+import type { search } from "../../search";
 import { searchRouter } from "../src/routers/search";
-import type { byTerm } from "../src/routers/search/byTerm";
 import { createTestContext } from "./helpers";
 
-type Input = inferProcedureInput<typeof byTerm>;
+type Input = inferProcedureInput<typeof search>;
 
 describe("Search route", async () => {
   let searchCaller: ReturnType<typeof searchRouter.createCaller>;
@@ -18,7 +18,7 @@ describe("Search route", async () => {
   describe("byTerm", () => {
     it("should search by address", async () => {
       const input: Input = {
-        term: "0xad01b55d7c3448b8899862eb335fbb17075d8de2",
+        query: "0xad01b55d7c3448b8899862eb335fbb17075d8de2",
       };
 
       const result = await searchCaller.byTerm(input);
@@ -29,7 +29,8 @@ describe("Search route", async () => {
 
     it("should search by commitment", async () => {
       const input: Input = {
-        term: "0xb4f67eb0771fbbf1b06b88ce0e23383daf994320508d44dd30dbd507f598c0d9b3da5a152e41a0428375060c3803b983",
+        query:
+          "0xb4f67eb0771fbbf1b06b88ce0e23383daf994320508d44dd30dbd507f598c0d9b3da5a152e41a0428375060c3803b983",
       };
 
       const result = await searchCaller.byTerm(input);
@@ -44,7 +45,8 @@ describe("Search route", async () => {
 
     it("should search by blob hash", async () => {
       const input: Input = {
-        term: "0x010001c79d78a76fb9b4bab3896ee3ea32f3e2607da7801eb1a92da39d6c1368",
+        query:
+          "0x010001c79d78a76fb9b4bab3896ee3ea32f3e2607da7801eb1a92da39d6c1368",
       };
 
       const result = await searchCaller.byTerm(input);
@@ -59,7 +61,8 @@ describe("Search route", async () => {
 
     it("should search by tx hash", async () => {
       const input: Input = {
-        term: "0x5be77167b05f39ea8950f11b0da2bdfec6e04055030068b051ac5a43aaf251e9",
+        query:
+          "0x5be77167b05f39ea8950f11b0da2bdfec6e04055030068b051ac5a43aaf251e9",
       };
 
       const result = await searchCaller.byTerm(input);
@@ -74,7 +77,7 @@ describe("Search route", async () => {
 
     it("should search by block number", async () => {
       const input: Input = {
-        term: "1001",
+        query: "1001",
       };
 
       const result = await searchCaller.byTerm(input);
@@ -91,7 +94,7 @@ describe("Search route", async () => {
 
     it("should search by slot number", async () => {
       const input: Input = {
-        term: "101",
+        query: "101",
       };
 
       const result = await searchCaller.byTerm(input);
@@ -108,7 +111,7 @@ describe("Search route", async () => {
 
     it("should return empty for unknown term", async () => {
       const input: Input = {
-        term: "unknown",
+        query: "unknown",
       };
 
       const result = await searchCaller.byTerm(input);
